Revertendo... Merged revision(s) 35305 from trunk2: -- ACBrNFe.EnvEvento [*] Ajusta na rotina LerFromIni o campo idLote para o tipo correto.    Por: Diego Folieni ........
git-svn-id: svn://svn.code.sf.net/p/acbr/code/trunk2@35306 6e92efe7-b92a-0410-a9ec-f9e4e41bb3a6

+++ Fontes/ACBrDFe/ACBrNFe/Base/Servicos/ACBrNFe.EnvEvento.pas
@@ -1143,7 +1143,7 @@ begin
   INIRec := TMemIniFile.Create('');
   try
     LerIniArquivoOuString(AIniString, INIRec);
-    idLote := INIRec.ReadInt64( 'EVENTO', 'idLote',
+    idLote := INIRec.ReadInteger( 'EVENTO', 'idLote',
                                         INIRec.ReadInteger('CCE', 'idLote', 0));
 
     I := 1;
